deleted service-manager file as it is not used anymore and moved the types directly inside the service-manager package
