+++ apps/web/integrations/cosmos/pagination/index.ts
@@ -1,7 +1,11 @@
 import { createLoader, EngineConfigMetadata } from "@modularcloud/ecs";
 import { createModularCloud } from "@modularcloud/sdk";
 import { z } from "zod";
-import { JSONRPCResponse, Transaction, TxSearch } from "~/lib/service-manager";
+import {
+  JSONRPCResponse,
+  Transaction,
+  TxSearch,
+} from "service-manager/types/rpc.type";
 import { PaginationTransform } from "./pagination";
 
 export async function PaginationExtract(
