+++ apps/web/ui/tabs/index.tsx
@@ -1,4 +1,3 @@
-import Link from "next/link";
 import { PageArchetype } from "~/ecs/archetypes/page";
 import { asyncUseEntity } from "~/ecs/hooks/use-entity/server";
 import { FetchLoadArgs, slugify } from "~/lib/utils";
